requirements.txt
beautifulsoup4>=4.12
requests>=2.31
schedule>=1.2
python-telegram-bot>=21.0
python-dotenv>=0.19